services/probability-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pandas==2.2.3
numpy==2.1.1
scipy==1.14.1
python-dotenv==1.0.1
redis==5.1.1
